+++ test/browser/runner-legacy.js
@@ -1,4 +1,4 @@
-less.strictMaths = false;
+less.strictMath = false;
 less.strictUnits = false;
 
 describe("less.js legacy tests", function() {